requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pydantic
python-dotenv
pyyaml
